Clean and refactor tests.
