+++ Timeline.Tests/Helpers/TestClock.cs
@@ -1,0 +1,15 @@
+using System;
+using Timeline.Services;
+
+namespace Timeline.Tests.Helpers
+{
+    public class TestClock : IClock
+    {
+        public DateTime? MockCurrentTime { get; set; } = null;
+
+        public DateTime GetCurrentTime()
+        {
+            return MockCurrentTime.GetValueOrDefault(DateTime.Now);
+        }
+    }
+}
