+++ Timeline.Tests/Services/UserDetailServiceTest.cs
@@ -7,7 +7,6 @@ using System.Threading.Tasks;
 using Timeline.Entities;
 using Timeline.Services;
 using Timeline.Tests.Helpers;
-using Timeline.Tests.Mock.Data;
 using Xunit;
 
 namespace Timeline.Tests.Services
