+++ Timeline.Tests/UsernameValidatorUnitTest.cs
@@ -1,5 +1,6 @@
 ﻿using FluentAssertions;
 using Timeline.Models.Validation;
+using Timeline.Tests.Helpers;
 using Xunit;
 
 namespace Timeline.Tests
